Move helpers to ristretto.utils.py
Move the following utility functions to top level utils.py
- conjugate_transpose
- _get_sdist_func -> get_sdist_func
- nmf_data

Consequently remove ristretto.util module.
