+++ ristretto/__init__.py
@@ -8,17 +8,17 @@ from ristretto.mf.rlu import rlu
 from ristretto.mf.rqb import rqb
 from ristretto.mf.reigen import reigh, reigh_nystroem, reigh_nystroem_col
 
-
 # Dynamic Mode Decomposition
 from ristretto.dmd.dmd import dmd
 from ristretto.dmd.rdmd import rdmd
 from ristretto.dmd.rdmd_single import rdmd_single
 
-
 #Nonnegative MF
 from ristretto.nmf.nmf_fhals import nmf
 from ristretto.nmf.nmf_fhals import rnmf
 
-
 #PCA
 from ristretto.pca.spca import spca, rspca, robspca
+
+#Utilities
+from ristretto import utils
