+++ ristretto/mf/interp_decomp.py
@@ -10,8 +10,7 @@ from __future__ import division
 import numpy as np
 from scipy import linalg
 
-from ..dmd.utils import conjugate_transpose
-from ..dmd.rdmd import _get_sdist_func
+from ..utils import conjugate_transpose, get_sdist_func
 
 _VALID_DTYPES = (np.float32, np.float64, np.complex64, np.complex128)
 _VALID_SDISTS = ('uniform', 'normal')
@@ -211,7 +210,7 @@ def rinterp_decomp(A, k=None, mode='column', p=10, q=1, sdist='normal', index_se
         raise ValueError("Target rank k must be >= 1 or < min(m, n), not %d" % k)
 
     # distribution to draw random samples
-    sdist_func = _get_sdist_func(sdist)
+    sdist_func = get_sdist_func(sdist)
 
     #Generate a random test matrix Omega
     Omega = sdist_func(size=(k+p, m)).astype(A.dtype)
@@ -353,7 +352,7 @@ def rinterp_decomp_qb(A, k=None, mode='column', p=10, q=1, sdist='normal', index
         raise ValueError("Target rank k must be >= 1 or < min(m, n), not %d" % k)
 
     # distribution to draw random samples
-    sdist_func = _get_sdist_func(sdist)
+    sdist_func = get_sdist_func(sdist)
 
     #Generate a random test matrix Omega
     Omega = sdist_func(size=(n, k+p)).astype(A.dtype)
