+++ ristretto/mf/reigen.py
@@ -11,8 +11,7 @@ import numpy as np
 from scipy import linalg
 from scipy.sparse import linalg as splinalg
 
-from ..dmd.utils import conjugate_transpose
-from ..dmd.rdmd import _get_sdist_func
+from ..utils import conjugate_transpose, get_sdist_func
 
 _VALID_DTYPES = (np.float32, np.float64, np.complex64, np.complex128)
 _VALID_SDISTS = ('uniform', 'normal')
@@ -80,7 +79,7 @@ def reigh(A, k, p=20, q=2, sdist='normal'):
         raise ValueError("Target rank k must be >= 1 or < min(m, n), not %d" % k)
 
     # distribution to draw random samples
-    sdist_func = _get_sdist_func(sdist)
+    sdist_func = get_sdist_func(sdist)
 
     #Generate a random test matrix Omega
     Omega = sdist_func(size=(n, k+p)).astype(A.dtype)
@@ -183,7 +182,7 @@ def reigh_nystroem(A, k, p=10, q=2, sdist='normal'):
         raise ValueError("Target rank k must be >= 1 or < min(m, n), not %d" % k)
 
     # distribution to draw random samples
-    sdist_func = _get_sdist_func(sdist)
+    sdist_func = get_sdist_func(sdist)
 
     #Generate a random test matrix Omega
     Omega = sdist_func(size=(n, k+p)).astype(A.dtype)
