--- ristretto/util/__init__.py
@@ -1,3 +1,0 @@
-from .data import nmf_data
-
-__all__ = ['nmf_data']
